create responseGoogle function and and index file to make the code cleaner
