+++ src/utils/emailValidation.js
@@ -1,4 +1,4 @@
-function validateEmail(email) {
+export function validateEmail(email) {
   const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
   return emailRegex.test(email);
 }
