+++ src/utils/index.js
@@ -1,0 +1,3 @@
+export { responsGoogleFunction } from "./responsGoogleFunction";
+export { validatePass } from "./passwordValidation";
+export { validateEmail } from "./emailValidation";
