+++ src/utils/passwordValidation.js
@@ -1,4 +1,4 @@
-function validatePass(password) {
+export function validatePass(password) {
   if (password.length >= 8) {
     return true;
   }
